requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-autofixture==0.12.1
django-celery-beat==2.0.0
django-celery-results==1.1.2
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-extensions==2.2.6
django-filter==2.2.0
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt-custom-user==1.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
idna==2.8
importlib-metadata==1.2.0
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.6
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
openapi-codec==1.3.2
overrides==2.8.0
packaging==20.3
Pillow==6.2.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
<<<<<<< HEAD
pyparsing==2.4.6
=======
python-crontab==2.4.0
python-dateutil==2.8.1
>>>>>>> 281635d50afe346479a16c663b91eb8226b6a9d1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
wrapt==1.11.2
zipp==0.6.0
